+++ src/components/agents/AgentUploadDialog.vue
@@ -1,74 +1,30 @@
 <template>
   <v-dialog
-    ref="uploadDialog"
+    ref="agentUploadDialog"
     v-model="show"
     max-width="750px"
   >
     <v-card>
       <v-card-title>
-        <span class="headline">Upload</span>
+        <span class="headline">Upload To Agent</span>
       </v-card-title>
       <v-card-text>
-        <v-form
-          ref="form"
-          on-submit="return false;"
-          @submit.prevent
-        >
-          <v-container>
-            <v-row>
-              <v-col cols="12">
-                <v-radio-group
-                  v-model="uploadType"
-                  mandatory
-                  row
-                >
-                  <v-radio
-                    value="uploadnew"
-                    label="Upload New File"
-                  />
-                  <v-radio
-                    value="useexisting"
-                    label="Use Existing File"
-                  />
-                </v-radio-group>
-                <v-file-input
-                  v-if="uploadType === 'uploadnew'"
-                  ref="fileInput"
-                  v-model="file"
-                  accept="*/*"
-                  :rules="rules['fileInput']"
-                />
-                <v-autocomplete
-                  v-else
-                  v-model="selectedDownload"
-                  :items="items"
-                  :loading="isLoading"
-                  :search-input.sync="search"
-                  :rules="rules['selectedDownload']"
-                  hide-no-data
-                  hide-selected
-                  cache-items
-                  item-text="description"
-                  item-value="filename"
-                  label="Server Files"
-                  placeholder="Start typing to Search"
-                  prepend-icon="mdi-database-search"
-                  return-object
-                  outlined
-                  dense
-                />
-                <v-text-field
-                  v-model="internalPathToFile"
-                  label="path/to/file (On the agent's machine)"
-                  :rules="rules['pathToFile']"
-                  outlined
-                  dense
-                  required
-                />
-              </v-col>
-            </v-row>
-          </v-container>
-        </v-form>
+        <p>Only showing files smaller than {{ formatBytes(maxBytes) }}</p>
+        <file-input
+          v-model="file"
+          :rules="rules['fileInput']"
+          :maximum-file-size="maxBytes"
+          return-object
+        />
+        <v-text-field
+          v-model="internalPathToFile"
+          label="path/to/file (On the agent's machine)"
+          :rules="rules['pathToFile']"
+          outlined
+          dense
+          required
+          :disabled="!file"
+        />
       </v-card-text>
       <v-card-actions>
         <v-spacer />
@@ -83,9 +39,10 @@
           color="blue darken-1"
           text
           :loading="loading"
+          :disabled="submitDisabled"
           @click="submit"
         >
-          Save
+          Upload
         </v-btn>
       </v-card-actions>
     </v-card>
@@ -93,11 +50,13 @@
 </template>
 
 <script>
-import * as downloadApi from '@/api/download-api';
+import FileInput from '@/components/FileInput.vue';
+import formatBytes from '@/utils/format-bytes';
 
 const MAX_BYTES = 1048576;
 
 export default {
+  components: { FileInput },
   props: {
     value: Boolean,
     language: {
@@ -115,11 +74,9 @@ export default {
   },
   data() {
     return {
-      uploadType: 'uploadnew',
-      isLoading: false,
+      formatBytes,
+      maxBytes: MAX_BYTES,
       descriptionLimit: 80,
-      selectedDownload: {},
-      search: null,
       entries: [],
       internalPathToFile: this.pathToFile,
       file: null,
@@ -128,10 +85,6 @@ export default {
           (v) => !!v || 'PathToFile is required',
         ],
         fileInput: [
-          (v) => !!v || 'File required',
-          (v) => (!!v && v.size < MAX_BYTES) || `Maximum size of ${Math.floor(MAX_BYTES / 1e6)} MiB.`,
-        ],
-        selectedDownload: [
           (v) => !!v || 'File required',
           (v) => (!!v && v.size < MAX_BYTES) || `Maximum size of ${Math.floor(MAX_BYTES / 1e6)} MiB.`,
         ],
@@ -148,16 +101,13 @@ export default {
       },
     },
     fileName() {
-      return this.file ? this.file.name : '';
+      return this.file ? this.file.filename : '';
     },
-    items() {
-      return this.entries.map((entry) => {
-        const description = entry.location.length > this.descriptionLimit
-          ? `(${this.formatBytes(entry.size)}) ${entry.location.slice(0, this.descriptionLimit)}...`
-          : `(${this.formatBytes(entry.size)}) ${entry.location}`;
-
-        return { ...entry, description };
-      });
+    fileId() {
+      return this.file ? this.file.id : '';
+    },
+    submitDisabled() {
+      return !this.file || !this.internalPathToFile || this.loading;
     },
   },
   watch: {
@@ -166,103 +116,43 @@ export default {
     },
     value(val) {
       if (val === false) {
-        this.$refs.form.reset();
+        this.file = null;
+        this.internalPathToFile = null;
       }
     },
-    file(val) {
+    fileName(val) {
       if (val) {
-        if (this.language === 'python') {
-          if (!this.internalPathToFile || this.internalPathToFile.length === 0) {
-            this.internalPathToFile = `/tmp/${val.name}`;
-          } else if (this.internalPathToFile.endsWith('/')) {
-            this.internalPathToFile += val.name;
+        if (['python', 'ironpython'].includes(this.language.toLowerCase())) {
+          // always use the passed in pathToFile if it exists
+          if (this.pathToFile) {
+            this.internalPathToFile = this.addTrailingSlash(this.pathToFile) + val;
           } else {
-            this.internalPathToFile = `/tmp/${val.name}`;
+            this.internalPathToFile = `/tmp/${val}`;
           }
-        } else if (this.language === 'powershell') {
-          if (!this.internalPathToFile || this.internalPathToFile.length === 0) {
-            this.internalPathToFile = `C:\\tmp\\${val.name}`;
-          } else if (this.internalPathToFile.endsWith('/') || this.internalPathToFile.endsWith('\\')) {
-            this.internalPathToFile += val.name;
+        } else if (['powershell', 'csharp'].includes(this.language.toLowerCase())) {
+          if (this.pathToFile) {
+            this.internalPathToFile = this.addTrailingSlash(this.pathToFile) + val;
           } else {
-            this.internalPathToFile = `C:\\tmp\\${val.name}`;
+            this.internalPathToFile = `C:\\tmp\\${val}`;
           }
         }
       }
-    },
-    selectedDownload(val) {
-      if (val) {
-        const { filename } = val;
-        if (this.language === 'python') {
-          if (!this.internalPathToFile || this.internalPathToFile.length === 0) {
-            this.internalPathToFile = `/tmp/${filename}`;
-          } else if (this.internalPathToFile.endsWith('/')) {
-            this.internalPathToFile += filename;
-          } else {
-            this.internalPathToFile = `/tmp/${filename}`;
-          }
-        } else if (this.language === 'powershell') {
-          if (!this.internalPathToFile || this.internalPathToFile.length === 0) {
-            this.internalPathToFile = `C:\\tmp\\${filename}`;
-          } else if (this.internalPathToFile.endsWith('/') || this.internalPathToFile.endsWith('\\')) {
-            this.internalPathToFile += filename;
-          } else {
-            this.internalPathToFile = `C:\\tmp\\${filename}`;
-          }
-        }
-      }
-    },
-    search() {
-      // For now we're just ignoring the search value and loading all the files.
-      // If we find that this is slow we can make the autocomplete more complicated.
-      // Items have already been loaded
-      if (this.items.length > 0) return;
-
-      // Items have already been requested
-      if (this.isLoading) return;
-
-      this.isLoading = true;
-
-      // Lazily load input items
-      downloadApi.getDownloads({ page: 1, limit: -1 })
-        .then((res) => {
-          const { total, records } = res;
-          this.count = total;
-          this.entries = records;
-        })
-        .catch((err) => {
-          console.log(err);
-        })
-        .finally(() => { this.isLoading = false; });
     },
   },
   methods: {
-    async submit() {
-      if (!this.$refs.form.validate()) { return; }
-
-      // todo if uploading a new file, do it here and then emit
-      let fileId = null;
-      if (this.uploadType === 'uploadnew') {
-        const formData = new FormData();
-        formData.append('file', this.file);
-
-        const response = await downloadApi.createDownload(formData);
-        fileId = response.data.id;
-      } else {
-        fileId = this.selectedDownload.id;
+    addTrailingSlash(val) {
+      if (val.endsWith('/') || val.endsWith('\\')) {
+        return val;
+      }
+      if (['python', 'ironpython'].includes(this.language.toLowerCase())) {
+        return `${val}/`;
       }
 
-      // otherwise emit the file id
-      this.$emit('submit', { file: fileId, pathToFile: this.internalPathToFile });
+      // todo need to test csharp and powershell agents for slash types
+      return `${val}\\`;
     },
-    // https://gist.github.com/zentala/1e6f72438796d74531803cc3833c039c
-    formatBytes(bytes, decimals) {
-      if (bytes === 0) return '0 Bytes';
-      const k = 1024;
-      const dm = decimals || 2;
-      const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-      const i = Math.floor(Math.log(bytes) / Math.log(k));
-      return `${parseFloat((bytes / (k ** i)).toFixed(dm))} ${sizes[i]}`;
+    async submit() {
+      this.$emit('submit', { file: this.fileId, pathToFile: this.internalPathToFile });
     },
   },
 };